3.4-django-testing/django_testing/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
model-bakery==1.15.0
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
